# 8314850: SharedRuntime::handle_wrong_method() gets called too often when resolving Continuation.enter