@olivergillespie 2023-08-23
Thanks! Indeed lowering `jdk.virtualThreadScheduler.parallelism` helps a lot and the improvement of the patch compared to before is quite clear at lower values. I'll be sure to test out the patch in my real workloads later.